Fixed problems from clang-tidy

+++ src/Utilities/Timer.cpp
@@ -5,7 +5,7 @@ Timer::Timer() {
     Reset();
 }
 
-long long Timer::GetTime() {
+int64_t Timer::GetTime() {
     auto elapsed = std::chrono::high_resolution_clock::now() - start_;
     return std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count();
 }
